Renamed requests.routes.js to connectRequest.routes.js and refactored to maintain naming convention.
